django-api/requirements.txt
Django>=3.2
redis
requests
psycopg2-binary>=2.8
django-celery-beat>=2.2.0
djangorestframework
django-filter
django-cors-headers
torch 
torchvision
opencv-contrib-python
pretrainedmodels
celery==5.1.2